Ask if neovim config should be installed

+++ install.sh
@@ -14,13 +14,18 @@ cat <<EOF > "$HOME/.gitconfig"
     path = $PWD/gitconfig
 EOF
 
-echo "Linking .vimrc ..."
-rm -f "$HOME/.vimrc"
-ln -s "$PWD/init.vim" "$HOME/.vimrc"
-mkdir "$HOME/.config/nvim"
-ln -s "$PWD/init.vim" "$HOME/.config/nvim/init.vim"
-
 echo "Linking .tmux.conf ..."
 rm -f "$HOME/.tmux.conf"
 ln -s "$PWD/tmux.conf" "$HOME/.tmux.conf"
+
+echo "Linking .vimrc ..."
+rm -f "$HOME/.vimrc"
+ln -s "$PWD/init.vim" "$HOME/.vimrc"
+
+read -p "neovim? "
+if [[ $REPLY = "y" ]]; then
+    echo "Linking init.vim ..."
+    mkdir -p "$HOME/.config/nvim"
+    ln -s "$PWD/init.vim" "$HOME/.config/nvim/init.vim"
+fi
